fix: component file directory
