+++ .eslintignore
@@ -6,5 +6,5 @@ coverage
 *.scss
 *.woff
 *.ttf
-src/index.ts
+packages/index.ts
 dist
