+++ index.html
@@ -8,6 +8,6 @@
 </head>
 <body>
     <div id="root"></div>
-    <script type="module" src="/src/index.tsx"></script>
+    <script type="module" src="/packages/index.tsx"></script>
 </body>
 </html>
