+++ packages/components/Button/index.tsx
@@ -1,50 +1,7 @@
 import React, { Component } from "react";
 import Icon from "../Icon";
+import {ISize, IColor, IIcon, ButtonProps } from './props';
 
-type ISize = "small" | "medium" | "large";
-type IColor =
-  | ""
-  | "black"
-  | "gray"
-  | "red"
-  | "yellow"
-  | "green"
-  | "blue"
-  | "indigo"
-  | "purple"
-  | "pink"
-  | "white";
-type IIcon =
-  | ""
-  | "juzhongduiqi"
-  | "zuuoduiqi"
-  | "yiwen"
-  | "xuanzewendnag"
-  | "youduiqi"
-  | "xunhuan"
-  | "bianji"
-  | "xiugai"
-  | "xinhao"
-  | "xiaoxi"
-  | "xiazai2"
-  | "tianjiawenjian"
-  | "tianjiawendang"
-  | "tianjia2"
-  | "tianjia1"
-  | "tixing"
-  | "tishi"
-  | "suoxiao"
-  | "gongzuotai"
-  | "zhichuhetong";
-
-interface ButtonProps {
-  color?: IColor;
-  icon?: IIcon;
-  size?: ISize;
-  round?: false | true;
-  plain?: false | true;
-  children?: string;
-}
 const AButton = (props: ButtonProps) => {
   const sizeOptions: Record<string, Record<string, string>> = {
     small: {
@@ -107,4 +64,5 @@ AButton.defaultProps = {
   round: false,
   plain: false,
 };
+
 export default AButton;
