+++ packages/components/Icon/props.ts
@@ -1,0 +1,9 @@
+interface iconProps {
+    iconName: string | undefined;
+    customClassName?: string;
+    onIconClick?: () => void;
+}
+
+export {
+    iconProps
+}
