+++ packages/entry.ts
@@ -1,0 +1,7 @@
+import AButton from "./components/Button";
+import AIcon from "./components/Icon";
+
+export { 
+    AButton,
+    AIcon
+ };
