--- src/Button/entry.ts
@@ -1,3 +1,0 @@
-import AButton from "./index";
-// 导出单独组件
-export { AButton };
