--- src/entry.ts
@@ -1,2 +1,0 @@
-import AButton from "./Button";
-export { AButton };
